+++ apps/carp_mobile_sensing_app/lib/main.dart
@@ -52,7 +52,7 @@ void main() async {
   await bloc.initialize(
     deploymentMode: DeploymentMode.local,
     // deploymentId: testDeploymentId,
-    // useCachedStudyDeployment: false,
+    useCachedStudyDeployment: false,
     resumeSensingOnStartup: false,
   );